Migrate to Open-Next and update Notion SDK
This version migrates to open-next and supports the wrangler runtime
fetch. To fix the bug induced by the migration, I found the solution was
to update the notion SDK to version 2025-09-03. Besides, to be
compatible to this version, I also changed the query to database to be
the query to data source.